feat(BPAAS-2137): continue startup upon opening min number of peer connections

+++ cmd/arc/services/metamorph.go
@@ -364,7 +364,7 @@ func setupMtmBcNetworkCommunication(l *slog.Logger, s store.MetamorphStore, arcC
 
 	// wait until min peer connections are ready and then continue startup while remaining peers connect
 	<-connectionsReady
-	l.Info("current open peer connections", slog.Int("count", int(manager.CountConnectedPeers())))
+	l.Info("current open peer connections", slog.Uint64("count", uint64(manager.CountConnectedPeers())))
 
 	// connect to mcast
 	if cfg.Mode == "hybrid" {
